- fix capital casing bug in PricingUtils - remove stupid AI comments
