+++ components/Premium/BuySubscription/BuySubscription.tsx
@@ -14,10 +14,8 @@ import {
     getProviderOriginalPrice,
     getProviderCurrencyCode,
     getDiscountPercent,
-    getTierApiProductId,
-    getTierProductId,
-    getFallbackSubscriptionPrice
-} from '../../../utils/pricingUtils'
+    getTierApiProductId
+} from '../../../utils/PricingUtils'
 
 interface Props {
     activePremiumProduct: PremiumProduct
@@ -137,17 +135,17 @@ function BuySubscription(props: Props) {
     // If we have wizard selections, use them to determine the selected type and duration
     const wizardSelectedType = props.selectedTier
         ? PREMIUM_TYPES.find(type => {
-              switch (props.selectedTier) {
-                  case PremiumTier.PREMIUM:
-                      return type.productId === 'premium'
-                  case PremiumTier.PREMIUM_PLUS:
-                      return type.productId === 'premium_plus'
-                  case PremiumTier.STARTER:
-                      return type.productId === 'starter_premium'
-                  default:
-                      return type.productId === 'premium'
-              }
-          })
+            switch (props.selectedTier) {
+                case PremiumTier.PREMIUM:
+                    return type.productId === 'premium'
+                case PremiumTier.PREMIUM_PLUS:
+                    return type.productId === 'premium_plus'
+                case PremiumTier.STARTER:
+                    return type.productId === 'starter_premium'
+                default:
+                    return type.productId === 'premium'
+            }
+        })
         : undefined
 
     const wizardIsYearOption = props.selectedDuration === Duration.YEARLY
@@ -231,9 +229,8 @@ function BuySubscription(props: Props) {
                     <p>
                         <strong>Tier:</strong>{' '}
                         <span
-                            className={`${styles.summaryValue} ${props.selectedTier === PremiumTier.PREMIUM ? styles.tierPremium : ''} ${
-                                props.selectedTier === PremiumTier.PREMIUM_PLUS ? styles.tierPremiumPlus : ''
-                            }`}
+                            className={`${styles.summaryValue} ${props.selectedTier === PremiumTier.PREMIUM ? styles.tierPremium : ''} ${props.selectedTier === PremiumTier.PREMIUM_PLUS ? styles.tierPremiumPlus : ''
+                                }`}
                         >
                             {getDisplayTierName()}
                         </span>
@@ -347,8 +344,8 @@ function BuySubscription(props: Props) {
                         )}
                         {appliedCreatorCode && !pricingError && (
                             <div style={{ marginTop: '10px', color: getDiscountPercentValue(getProductIdForTier(props.selectedTier!, yearOption)) ? 'green' : 'blue' }}>
-                                {getDiscountPercentValue(getProductIdForTier(props.selectedTier!, yearOption)) 
-                                    ? `Creator code applied! You get ${getDiscountPercentValue(getProductIdForTier(props.selectedTier!, yearOption))}% off` 
+                                {getDiscountPercentValue(getProductIdForTier(props.selectedTier!, yearOption))
+                                    ? `Creator code applied! You get ${getDiscountPercentValue(getProductIdForTier(props.selectedTier!, yearOption))}% off`
                                     : 'Creator code applied'}
                             </div>
                         )}
@@ -370,7 +367,7 @@ function BuySubscription(props: Props) {
                             const activeEl = document.activeElement as HTMLElement | null
                             if (activeEl && activeEl.tagName === 'BUTTON') {
                                 activeEl.innerText = 'Redirecting to payment provider...'
-                                ;(activeEl as HTMLButtonElement).disabled = true
+                                    ; (activeEl as HTMLButtonElement).disabled = true
                             }
                         }}
                     >
