+++ utils/PricingUtils.tsx
@@ -1,3 +1,4 @@
+import { BatchProductPricingResponse, ProviderPricingOption } from '../api/_generated/skyApi.schemas'
 import { PremiumTier } from '../components/Premium/PremiumPurchaseWizard/types'
 
 // Base prices in EUR (before tax)
@@ -113,15 +114,101 @@ export function calculatePrice(tier: PremiumTier, countryCode?: string, discount
     }
 }
 
-export function getPricingPeriodText(tier: PremiumTier): string {
-    switch (tier) {
-        case PremiumTier.STARTER:
-            return 'per year'
-        case PremiumTier.PREMIUM:
-            return 'per month'
-        case PremiumTier.PREMIUM_PLUS:
-            return 'one-time'
-        default:
-            return ''
-    }
+export const CURRENCY_SYMBOLS: Record<string, string> = {
+    'USD': '$',
+    'EUR': '€',
+    'GBP': '£',
+    'INR': '₹',
+    'JPY': '¥',
+    'CNY': '¥',
+    'CAD': 'C$',
+    'AUD': 'A$'
+}
+
+export const getCurrencySymbol = (currencyCode: string): string => {
+    return CURRENCY_SYMBOLS[currencyCode] || currencyCode
+}
+
+export const getProvider = (
+    pricingData: BatchProductPricingResponse | null,
+    productSlug: string,
+    providerSlug: string
+): ProviderPricingOption | undefined => {
+    const product = pricingData?.products?.find(p => p.productSlug === productSlug)
+    return product?.providers?.find(p => p.providerSlug === providerSlug)
+}
+
+export const getProviderPrice = (
+    pricingData: BatchProductPricingResponse | null,
+    productSlug: string,
+    providerSlug: string
+): number | null => {
+    const provider = getProvider(pricingData, productSlug, providerSlug)
+    return provider ? (provider.discountedPrice ?? provider.originalPrice) : null
+}
+
+export const getProviderOriginalPrice = (
+    pricingData: BatchProductPricingResponse | null,
+    productSlug: string,
+    providerSlug: string
+): number | null => {
+    return getProvider(pricingData, productSlug, providerSlug)?.originalPrice ?? null
+}
+
+export const getProviderCurrencyCode = (
+    pricingData: BatchProductPricingResponse | null,
+    productSlug: string,
+    providerSlug: string
+): string => {
+    return getProvider(pricingData, productSlug, providerSlug)?.currencyCode ?? 'EUR'
+}
+
+export const getDiscountPercent = (
+    pricingData: BatchProductPricingResponse | null,
+    productSlug: string
+): number | null => {
+    return pricingData?.products?.find(p => p.productSlug === productSlug)?.discountPercent ?? null
+}
+
+const TIER_PRODUCT_MAP: Record<PremiumTier, string> = {
+    [PremiumTier.PREMIUM]: 'premium',
+    [PremiumTier.PREMIUM_PLUS]: 'premium_plus',
+    [PremiumTier.STARTER]: 'starter_premium'
+}
+
+export const getTierProductId = (tier: PremiumTier): string => {
+    return TIER_PRODUCT_MAP[tier] ?? 'premium'
+}
+
+const TIER_SLUG_MAP: Record<PremiumTier, { monthly: string; yearly: string }> = {
+    [PremiumTier.PREMIUM]: { monthly: 'l_premium', yearly: 'l_premium-year' },
+    [PremiumTier.PREMIUM_PLUS]: { monthly: 'l_prem_plus', yearly: 'l_prem_plus-year' },
+    [PremiumTier.STARTER]: { monthly: 'l_starter_premium', yearly: 'l_starter_premium' }
+}
+
+export const getTierSubscriptionSlug = (tier: PremiumTier, isYearly: boolean): string => {
+    const slugs = TIER_SLUG_MAP[tier]
+    return isYearly ? slugs.yearly : slugs.monthly
+}
+
+const TIER_API_PRODUCT_MAP: Record<PremiumTier, string> = {
+    [PremiumTier.PREMIUM]: 'l_premium',
+    [PremiumTier.PREMIUM_PLUS]: 'l_prem_plus',
+    [PremiumTier.STARTER]: 'l_starter_premium'
+}
+
+export const getTierApiProductId = (tier: PremiumTier, isYearly: boolean): string => {
+    const baseId = TIER_API_PRODUCT_MAP[tier]
+    return isYearly && tier !== PremiumTier.STARTER ? `${baseId}-year` : baseId
+}
+
+const SUBSCRIPTION_PRICES: Record<string, { monthly: number; yearly: number }> = {
+    'premium': { monthly: 8.69, yearly: 96.69 },
+    'premium_plus': { monthly: 35.69, yearly: 354.2 },
+    'starter_premium': { monthly: 16.99, yearly: 16.99 }
+}
+
+export const getFallbackSubscriptionPrice = (productId: string, isYearly: boolean): number => {
+    const prices = SUBSCRIPTION_PRICES[productId]
+    return prices ? (isYearly ? prices.yearly : prices.monthly) : -1
 }
